Show icon in edit mode with different type of wrapper, apply for view mode. Origin link for search utility in view mode
