+++ src/app/components/iconComponent.tsx
@@ -1,21 +1,7 @@
+import { getDomainUrl } from '@/app/lib/utilities';
 
-export default function IconComponent({size, url}: {size: number, url: string}) {
+export default function IconComponent({size, url, clsName}: {size: number, url: string, clsName: string}) {
     
-    function getDomainUrl(url: string) {
-        const fIndex = 8;
-        
-        let lIndex = url.indexOf('/', fIndex);
-        if (lIndex < 0) {
-            lIndex = url.indexOf('?', fIndex);
-            if (lIndex < 0) {
-                lIndex = url.length;
-            }
-        }
-        
-        const domainUrl = url.substring(0, lIndex);
-        return domainUrl;
-    }
-
     function getFavicon(size: number, domainUrl: string) {
         const requestGoogle = `https://www.google.com/s2/favicons?sz=${size}&domain_url=${domainUrl}`;
         return requestGoogle;
@@ -26,7 +12,7 @@ export default function IconComponent({size, url}: {size: number, url: string}) 
 
     return (
         <>
-            <img src={iconUrl} width={size} height={size} alt={''} className="iconWrapper"/>&nbsp;
+            <img src={iconUrl} width={size} height={size} alt={''} className={clsName}/>&nbsp;
         </>
     );
 }
